+++ cnccode/controller_classes/CTHeader.inc.php
@@ -24,6 +24,7 @@ define(
 class CTHeader extends CTCNC
 {
     const GET_PRIORITIES_DESCRIPTIONS = 'getPrioritiesDescriptions';
+    const GET_NUMBER_ALLOWED_MISTAKES="numberOfAllwoedMistaks";
     /** @var DSForm */
     public $dsHeader;
     /** @var BUHeader */
@@ -71,6 +72,9 @@ class CTHeader extends CTCNC
                 $this->getPrioritiesDescriptions();
                 exit;
             }
+            case self::GET_NUMBER_ALLOWED_MISTAKES:
+                echo json_encode($this->getNumberAllwoedMistakes());
+                exit;
             case CTHEADER_ACT_EDIT:
             default:
                 $this->checkPermissions(MAINTENANCE_PERMISSION);
@@ -654,6 +658,7 @@ class CTHeader extends CTCNC
                 DBEHeader::holdAllSOProjectsP5sforQAReview                           => $dsHeader->getValue(
                     DBEHeader::holdAllSOProjectsP5sforQAReview
                 ) ? "checked" : null,
+                DBEHeader::numberOfAllowedMistakes                                   => $dsHeader->getValue(DBEHeader::numberOfAllowedMistakes) ,
             ]
         );
         // VAT code
@@ -791,5 +796,11 @@ class CTHeader extends CTCNC
             ]
         );
 
+    }
+    function getNumberAllwoedMistakes()
+    {
+        $dbeHeader = new DBEHeader($this);
+        $dbeHeader->getRow(1);
+        return ["value"=>$dbeHeader->getValue(DBEHeader::numberOfAllowedMistakes)];
     }
 }