+++ htdocs/components/services/ApiUrls.js
@@ -7,6 +7,6 @@ export default class ApiUrls{
  static sdDashboard="/SDManagerDashboard.php?action=";
  static FirstTimeFixReport="/FirstTimeFixReport.php?action=";
  static RequestDashboard="/RequestDashBoard.php?action=";
-
  static KPIReport="/KPIReport.php?action=";
+ static Header="/Header.php?action=";
 }